+++ google-cloud-datastore/src/main/java/com/google/cloud/datastore/Datastore.java
@@ -18,7 +18,7 @@ package com.google.cloud.datastore;
 
 import com.google.api.core.BetaApi;
 import com.google.cloud.Service;
-import com.google.cloud.datastore.models.QueryProfile;
+import com.google.cloud.datastore.models.ExplainOptions;
 import com.google.datastore.v1.TransactionOptions;
 import java.util.Iterator;
 import java.util.List;
@@ -465,9 +465,8 @@ public interface Datastore extends Service<DatastoreOptions>, DatastoreReaderWri
   <T> QueryResults<T> run(Query<T> query, ReadOption... options);
 
   /**
-   * Submits a {@link Query} with specified {@link
-   * com.google.cloud.datastore.models.QueryProfile.QueryMode} and returns its result. {@link
-   * ReadOption}s can be specified if desired.
+   * Submits a {@link Query} with specified {@link com.google.cloud.datastore.models.ExplainOptions}
+   * and returns its result. {@link ReadOption}s can be specified if desired.
    *
    * <p>Example of running a query to find all entities of one kind.
    *
@@ -476,15 +475,14 @@ public interface Datastore extends Service<DatastoreOptions>, DatastoreReaderWri
    * StructuredQuery<Entity> query = Query.newEntityQueryBuilder()
    *     .setKind(kind)
    *     .build();
-   * QueryResults<Entity> results = datastore.run(query, QueryMode.EXPLAIN_ANALYZE);
-   * ResultSetStats resultSetStats = results.getResultSetStats();
+   * QueryResults<Entity> results = datastore.run(query, ExplainOptions.newBuilder().setAnalyze(true).build());
    * }</pre>
    *
    * @throws DatastoreException upon failure
    */
   @BetaApi
   default <T> QueryResults<T> run(
-      Query<T> query, QueryProfile.QueryMode queryMode, ReadOption... options) {
+      Query<T> query, ExplainOptions explainOptions, ReadOption... options) {
     throw new UnsupportedOperationException("Not implemented.");
   }
 
@@ -537,8 +535,8 @@ public interface Datastore extends Service<DatastoreOptions>, DatastoreReaderWri
 
   /**
    * Submits a {@link AggregationQuery} with specified {@link
-   * com.google.cloud.datastore.models.QueryProfile.QueryMode} and returns {@link
-   * AggregationResults}. {@link ReadOption}s can be specified if desired.
+   * com.google.cloud.datastore.models.ExplainOptions} and returns {@link AggregationResults}.
+   * {@link ReadOption}s can be specified if desired.
    *
    * <p>Example of running an {@link AggregationQuery} to find the count of entities of one kind.
    *
@@ -552,8 +550,7 @@ public interface Datastore extends Service<DatastoreOptions>, DatastoreReaderWri
    *    .addAggregation(count().as("total_count"))
    *    .over(selectAllQuery)
    *    .build();
-   * AggregationResults aggregationResults = datastore.runAggregation(aggregationQuery, QueryMode.EXPLAIN_ANALYZE);
-   * ResultSetStats aggregationStats = aggregationResults.getResultSetStats();
+   * AggregationResults aggregationResults = datastore.runAggregation(aggregationQuery, ExplainOptions.newBuilder().setAnalyze(true).build());
    * }</pre>
    *
    * @throws DatastoreException upon failure
@@ -561,7 +558,7 @@ public interface Datastore extends Service<DatastoreOptions>, DatastoreReaderWri
    */
   @BetaApi
   default AggregationResults runAggregation(
-      AggregationQuery query, QueryProfile.QueryMode queryMode, ReadOption... options) {
+      AggregationQuery query, ExplainOptions explainOptions, ReadOption... options) {
     throw new UnsupportedOperationException("Not implemented.");
   }
 }