+++ google-cloud-datastore/src/main/java/com/google/cloud/datastore/DatastoreReader.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.datastore;
 
-import com.google.cloud.datastore.models.QueryProfile;
+import com.google.cloud.datastore.models.ExplainOptions;
 import java.util.Iterator;
 import java.util.List;
 
@@ -66,13 +66,11 @@ public interface DatastoreReader {
 
   /**
    * Submits a {@link AggregationQuery} with a specified {@link
-   * com.google.cloud.datastore.models.QueryProfile.QueryMode} and returns {@link
-   * AggregationResults}.
+   * com.google.cloud.datastore.models.ExplainOptions} and returns {@link AggregationResults}.
    *
    * @throws DatastoreException upon failure
    */
-  default AggregationResults runAggregation(
-      AggregationQuery query, QueryProfile.QueryMode queryMode) {
+  default AggregationResults runAggregation(AggregationQuery query, ExplainOptions explainOptions) {
     throw new UnsupportedOperationException("Not implemented.");
   }
 }