+++ google-cloud-datastore/src/main/java/com/google/cloud/datastore/execution/AggregationQueryExecutor.java
@@ -23,9 +23,8 @@ import com.google.cloud.datastore.ReadOption;
 import com.google.cloud.datastore.ReadOption.QueryConfig;
 import com.google.cloud.datastore.execution.request.AggregationQueryRequestProtoPreparer;
 import com.google.cloud.datastore.execution.response.AggregationQueryResponseTransformer;
-import com.google.cloud.datastore.models.QueryProfile;
+import com.google.cloud.datastore.models.ExplainOptions;
 import com.google.cloud.datastore.spi.v1.DatastoreRpc;
-import com.google.datastore.v1.QueryMode;
 import com.google.datastore.v1.RunAggregationQueryRequest;
 import com.google.datastore.v1.RunAggregationQueryResponse;
 import java.util.Arrays;
@@ -50,20 +49,23 @@ public class AggregationQueryExecutor
 
   @Override
   public AggregationResults execute(
-      AggregationQuery query, QueryProfile.QueryMode queryMode, ReadOption... readOptions) {
+      AggregationQuery query, ExplainOptions explainOptions, ReadOption... readOptions) {
     RunAggregationQueryRequest runAggregationQueryRequest =
-        getRunAggregationQueryRequest(query, queryMode.toPb(), readOptions);
+        getRunAggregationQueryRequest(
+            query, explainOptions == null ? null : explainOptions.toPb(), readOptions);
     RunAggregationQueryResponse runAggregationQueryResponse =
         this.datastoreRpc.runAggregationQuery(runAggregationQueryRequest);
     return this.responseTransformer.transform(runAggregationQueryResponse);
   }
 
   private RunAggregationQueryRequest getRunAggregationQueryRequest(
-      AggregationQuery query, QueryMode queryMode, ReadOption... readOptions) {
+      AggregationQuery query,
+      com.google.datastore.v1.ExplainOptions explainOptions,
+      ReadOption... readOptions) {
     QueryConfig<AggregationQuery> queryConfig =
         readOptions == null
-            ? QueryConfig.create(query, queryMode)
-            : QueryConfig.create(query, queryMode, Arrays.asList(readOptions));
+            ? QueryConfig.create(query, explainOptions)
+            : QueryConfig.create(query, explainOptions, Arrays.asList(readOptions));
     return this.protoPreparer.prepare(queryConfig);
   }
 }