+++ google-cloud-datastore/src/test/java/com/google/cloud/datastore/execution/response/AggregationQueryResponseTransformerTest.java
@@ -22,12 +22,15 @@ import static com.google.common.truth.Truth.assertThat;
 import com.google.cloud.Timestamp;
 import com.google.cloud.datastore.AggregationResult;
 import com.google.cloud.datastore.AggregationResults;
+import com.google.cloud.datastore.models.ExplainMetrics;
 import com.google.common.collect.ImmutableMap;
 import com.google.datastore.v1.AggregationResultBatch;
-import com.google.datastore.v1.QueryPlan;
-import com.google.datastore.v1.ResultSetStats;
+import com.google.datastore.v1.ExecutionStats;
+import com.google.datastore.v1.PlanSummary;
 import com.google.datastore.v1.RunAggregationQueryResponse;
 import com.google.datastore.v1.Value;
+import com.google.protobuf.Duration;
+import com.google.protobuf.Struct;
 import java.util.AbstractMap.SimpleEntry;
 import java.util.HashMap;
 import java.util.Map;
@@ -78,7 +81,7 @@ public class AggregationQueryResponseTransformerTest {
     assertThat(aggregationResults.get(0)).isEqualTo(new AggregationResult(toDomainValues(result1)));
     assertThat(aggregationResults.get(1)).isEqualTo(new AggregationResult(toDomainValues(result2)));
     assertThat(aggregationResults.getReadTime()).isEqualTo(readTime);
-    assertThat(aggregationResults.getResultSetStats().isPresent()).isFalse();
+    assertThat(aggregationResults.getExplainMetrics().isPresent()).isFalse();
   }
 
   @Test
@@ -108,10 +111,41 @@ public class AggregationQueryResponseTransformerTest {
                     .build())
             .setReadTime(readTime.toProto())
             .build();
-    ResultSetStats stats =
-        ResultSetStats.newBuilder().setQueryPlan(QueryPlan.newBuilder().build()).build();
+
+    ExecutionStats executionStats =
+        ExecutionStats.newBuilder()
+            .setDebugStats(
+                Struct.newBuilder()
+                    .putFields(
+                        "field",
+                        com.google.protobuf.Value.newBuilder().setStringValue("val").build())
+                    .build())
+            .setExecutionDuration(Duration.newBuilder().setSeconds(1).build())
+            .setReadOperations(1)
+            .setResultsReturned(2)
+            .build();
+
+    PlanSummary planSummary =
+        PlanSummary.newBuilder()
+            .addIndexesUsed(
+                Struct.newBuilder()
+                    .putFields(
+                        "field2",
+                        com.google.protobuf.Value.newBuilder().setStringValue("val2").build())
+                    .build())
+            .build();
+
+    com.google.datastore.v1.ExplainMetrics explainMetrics =
+        com.google.datastore.v1.ExplainMetrics.newBuilder()
+            .setExecutionStats(executionStats)
+            .setPlanSummary(planSummary)
+            .build();
+
     RunAggregationQueryResponse runAggregationQueryResponse =
-        RunAggregationQueryResponse.newBuilder().setBatch(resultBatch).setStats(stats).build();
+        RunAggregationQueryResponse.newBuilder()
+            .setBatch(resultBatch)
+            .setExplainMetrics(explainMetrics)
+            .build();
 
     AggregationResults aggregationResults =
         responseTransformer.transform(runAggregationQueryResponse);
@@ -120,8 +154,8 @@ public class AggregationQueryResponseTransformerTest {
     assertThat(aggregationResults.get(0)).isEqualTo(new AggregationResult(toDomainValues(result1)));
     assertThat(aggregationResults.get(1)).isEqualTo(new AggregationResult(toDomainValues(result2)));
     assertThat(aggregationResults.getReadTime()).isEqualTo(readTime);
-    assertThat(aggregationResults.getResultSetStats().get())
-        .isEqualTo(new com.google.cloud.datastore.models.ResultSetStats(stats));
+    assertThat(aggregationResults.getExplainMetrics().get())
+        .isEqualTo(new ExplainMetrics(explainMetrics));
   }
 
   @Test